Add enough register glue to make the thing read something that I think is the hardware revision.
The stuff in iwl/ is going to eventually be a combination of what's
from the openbsd driver that I've been given, which was based on the
linux iwlwifi driver.  I'm also going to keep it formatted like
the linux driver, so i hopefully can do directory diffs and figure
out what I'm missing / doing wrong.