Start fleshing out the firmware load stuff.
Doing the initial firmware load is actually quite complicated.
The firmware bit isn't too bad, but the notification path requires
the transmit and receive paths to be up, ready and working.

So the basics are here:

* get the hardware ids so we know which firmware we're looking for;
* populate sc_cfg with the config from above;
* wake up the hardware;
* this is done outside of any lock context so we don't sleep()
  inside the firmware stack;
* call iwa_prepare_card_hw() to set the initial things up;
* setup the queues;
* iwa_attach() loads in some firmware - for now it's hardcoded
  but eventually we'll iterate over the right firmware names
  until we find one that works (or doesn't);
* call iwa_preinit() wich kicks off the initial firmware parse
  and sets up the NIC with the firmware sections.

The initial firmware pass involves grabbing each section, loading
it into the NIC and waiting for an interrupt to say each section
is OK.  I'm getting these interrupts, so the firmware load seems
to work okay.

Then, once this is all done (and this is the bit that isn't done yet)
it'll send an RX notification to say "i'm ready".

The NVRAM/EEPROM bits aren't read until this point - including
the initial calibration details, TX/RX antenna configuration,
the channel list and MAC address.  That's needed by the rest
of the startup path.

At this point the transmit and receive queues are needed:

* the receive queue is needed so we can get RX notifications
  for things like calibration done, firmware load ready, firmware
  startup ready;
* the command queue is needed so we can complete the initial
  setup and calibration - it requires the driver to send
  an initial calibration command and configure the initial
  antennas.

So - I'm getting all the way down to the point where the initial
"hello i'm here!" RX notification is coming in - and I see the
notification interrupt happen, but I'm not yet checking the RX
queues.  That'll come next.

This code is a bit more of a mess than I'd like.  There's now
a /lot/ of code in the wrong files mainly so I can at least get
platform bringup completed before I move onto the next bits.

So what's next:

* implement enough of the RX notification path to receive
  completion notices and signal that we've finished startup
* read the board config from NVRAM and populate things;
* figure out how to implement sync and async command sending
  and response handling;
* finish the rest of the initial configuration startup pass.

Once that's done a big refactoring/reshuffle will need to occur.
I don't like how messy all of this is:

* The transaction layer should be a separate struct and not know
  about the softc at all (except perhaps a void pointer for
  debug echos);
* including iwavar.h and the softc requires a lot of the
  struct includes to be included, even if they're not required.
  I'll see what else can be separated out into separate
  opaque-ish structs;
* Separate out the command tx/rx code from the freebsd and busdma
  specific bits (busdma, mbufs, etc);
* unify the register access methods - maybe go back to iwa_{read/write}32
  rather than the freeBSD style.  Right now it's a bit too messy.

But holy crap, this bit works!