Add in a prototype for receive handling.
iwa_notif_intr() is from the openbsd driver, changed to work on freebsd.
Most of the commands are currently stubbed out, but the firmware completion
notification is in there.

This works enough to finish injecting the firmware into the chip and get
the MVM notification saying that the firmware load completed.

Next up - NVM (NVRAM/EEPROM) loading!