Re-implement the command wakeup bits.
This mirrors what's done in if_iwn - it sleeps on the descriptor
ring slot and will wake up when it's done.

This does time out after 5 seconds, rather than sleeping forever.
But it doesn't really know (yet) if the hardware has gone away
or the NIC was reset.  I'm going to have to handle that particular
gem later on, once all of this mess is finshed.

Next up - figuring out how to get the command response back
to the original caller, and then making sure everything is
freed correctly once it's done.