Ok, this is a mostly-comment commit, but I think it's worth describing what I've found in the driver(s) here.
The linux driver implements a blocking / synchronous command
interface as well as a way to get the reply data linked back
to the command that expects it.

However, it does some shady looking page stealing tricks behind
the scenes if the command indicated it wants the response
(WANT_SKB.)  This both means "I want the response data" and
"I'm going to be responsible for freeing it."

So, there seems to be some shady looking response path that is
moving the reply page(s) to the caller and, although I haven't found
it yet, it's replenishing them with new ones.

So the caller gets the buffers for the response.  It doesn't allocate
space to copy the response into - it just gets said response buffers.

The linux driver implements iwl_free_resp() as "free the underlying
pages that we stole back to the kernel."

The openbsd driver doesn't do this.  It's doing both a global
"want response" sleep/turnstile and it's also doing a per-descriptor
sleep for the response.  The global "want response" turnstile looks
like it's supposed to handle waiting for the response to complete -
it then will wakeup whatever's waiting for that.  But it's only doing
the sleep in iwm_send_cmd() - I don't really understand how the
receive / interrupt path is protected from concurrent packets being
received and churned through the ring whilst some thread is waiting
for the response wakeup so it can do its own thing.  the only thing
this is preventing is a second command coming along and wanting
a response - but I bet high enough data packet rates could cause
an overflow.

So, ugh.  I have to figure out what I'm going to do for the
FreeBSD version of this, as everything is SMP'-y and it's quite
possible we'll have this mess running concurrent at some point.

The cheat way is to allocate a new mbuf for that ring slot and
"gift" the mbuf to the transmit slot.  Then the response would
really just be an mbuf that's being freed in iwa_free_resp().
However, this means I'm going to be refilling mbufs in the
interrupt routine /and/ I have to hack up the iwl driver to
keep the mbuf pointer around as well as the iwl_rx_packet
reference (which is to the mbuf payload, not the mbuf itself.)

The cleaner, less efficient and more work way is to for now
require command responses to be copied into a buffer that the caller
allocates.  Commands won't happen anywhere near as often as
received packets, so we may as well just copy the frame out
and immediately release the mbuf back to the caller.  This requires
that I change all the routines to supply an actual response
buffer.

However it may not be that bad - I think once I implement the
actual RX packet path I'll end up recycling RX buffers and mbufs
anyway as that's what happens when RX slots get handed up to
the net80211 input side.  So, maybe I should just bite the
bullet and "steal" the mbuf away here.

Grr. Just, grr.

I should also have some value that gets set (or maybe just
a NULL pointer) if we fail to get the buffer during a command
response.  eg, if we try to allocate an mbuf to replace the
command slot with and we fail to do so.  I don't know what
the NIC will do if rx buffer slots are NULL.

Grrrr....