Implement a notification path back from the RX path through to the TX command slot back to the original iwl_host_cmd that initiated the command.
This is like Linux - there's a backpointer ("meta") that lets
the iwlwifi/trans code notify the original iwl_host_cmd of
the receive buffer with the contents its after.

This seems to work enough to not crash but I have no idea yet
whether the data that is showing up is correct.  It's certainly
failing the NVM read check.